sql: avoid extra cast to Datum
Previously, during comparison operator evaluation, we would always
re-cast some Datum values to Datum from Expr because of an detail around
how we look up the implementation of comparison operators.

Perhaps surprisingly, this showed up in profiles when a filter was
applied to a large amount of data, like a COUNT(*) with a filter on a
non-indexed column. This query is common and shows up in TPCC.

Release note: None