gossip: remove unnecessary gossip traffic
The previous handling of high water stamps left a gap during which an
info could be resent over a connection multiple times. The general
gossip send loop is to wait for a new info to be added to the local info
store, to compute a delta between the local info store and the remote
info store using the remote's high water stamps and to update the
remote's high water stamps when the remote sends a reply on the
connection. The problem here is that updating the remote's high water
stamps when the remote sends a reply leaves a window where an info added
to the local store would be sent multiple times (every time an
additional info was added) until the remote replied. The fix is to keep
track of a per connection high water stamp map and ratchet the contained
stamps both when we receive a reply from the remote and when we send the
remote an info. The result is a 30-40% reduction in gossip
traffic (bytes sent/received) as measured on idle 3, 9 and 27 node
clusters.

See #30088

Release note: None