sql: point spans should never include interleaves
Previously, many sql processors that needed to generate spans for key
lookups unconditionally used the PrefixEnd method on the key to look up
to get the end of the keyspan. While correct, this generated span is
wide enough to contain all child interleaves of a table on that key, if
they exist. This is very bad for performance, causing "point scans" to
accidentally need to scan through an entire interleaves if it exists.

Now, code is adjusted to use spans that are bounded on the right by the
interleave marker appended to the start key, if the index to look up in
is unique and fully specified. Note that the condition chosen is broader
than strictly necessary - we only insert interleaves today into primary
indexes, and we could tighten the condition to only emit an interleave
marker if the index does in fact have interleaves - but keeping the
condition simple makes things easier to test and provides a modicum of
assurance that this hairy code won't have to change again if we decide
to permit interleaving tables or indexes into other unique, non primary
indexes later.

New helper methods were introduced that permit retrieval of the key span
for a set of key values, and not just the key itself, in the hopes that
future code won't commit the same sins.

Release note (performance improvement): index joins, lookup joins,
foreign key checks, cascade scans, zig zag joins, and upserts no longer
needlessly scan over child interleaved tables when searching for keys.